refactor(core): clean up
