+++ wallet/wallet.go
@@ -8,8 +8,6 @@ import (
 	"github.com/golang/protobuf/proto"
 	"github.com/golang/protobuf/ptypes"
 	"github.com/op/go-logging"
-	cmodels "github.com/textileio/textile-go/central/models"
-	"github.com/textileio/textile-go/core/central"
 	"github.com/textileio/textile-go/crypto"
 	"github.com/textileio/textile-go/net"
 	nm "github.com/textileio/textile-go/net/model"
@@ -85,9 +83,7 @@ type Wallet struct {
 	pointerRepublisher *net.PointerRepublisher
 }
 
-const (
-	pingTimeout = time.Second * 10
-)
+const pingTimeout = time.Second * 10
 
 var ErrStarted = errors.New("node is already started")
 var ErrStopped = errors.New("node is already stopped")
@@ -300,96 +296,6 @@ func (w *Wallet) GetRepoPath() string {
 	return w.repoPath
 }
 
-// SignUp requests a new username and token from the central api and saves them locally
-func (w *Wallet) SignUp(reg *cmodels.Registration) error {
-	if err := w.touchDatastore(); err != nil {
-		return err
-	}
-	log.Debugf("signup: %s %s %s %s %s", reg.Username, "xxxxxx", reg.Identity.Type, reg.Identity.Value, reg.Referral)
-
-	// remote signup
-	res, err := central.SignUp(reg, w.GetCentralUserAPI())
-	if err != nil {
-		log.Errorf("signup error: %s", err)
-		return err
-	}
-	if res.Error != nil {
-		log.Errorf("signup error from central: %s", *res.Error)
-		return errors.New(*res.Error)
-	}
-
-	// local signin
-	if err := w.datastore.Profile().SignIn(
-		reg.Username,
-		res.Session.AccessToken, res.Session.RefreshToken,
-	); err != nil {
-		log.Errorf("local signin error: %s", err)
-		return err
-	}
-	return nil
-}
-
-// SignIn requests a token with a username from the central api and saves them locally
-func (w *Wallet) SignIn(creds *cmodels.Credentials) error {
-	if err := w.touchDatastore(); err != nil {
-		return err
-	}
-	log.Debugf("signin: %s %s", creds.Username, "xxxxxx")
-
-	// remote signin
-	res, err := central.SignIn(creds, w.GetCentralUserAPI())
-	if err != nil {
-		log.Errorf("signin error: %s", err)
-		return err
-	}
-	if res.Error != nil {
-		log.Errorf("signin error from central: %s", *res.Error)
-		return errors.New(*res.Error)
-	}
-
-	// local signin
-	if err := w.datastore.Profile().SignIn(
-		creds.Username,
-		res.Session.AccessToken, res.Session.RefreshToken,
-	); err != nil {
-		log.Errorf("local signin error: %s", err)
-		return err
-	}
-	return nil
-}
-
-// SignOut deletes the locally saved user info (username and tokens)
-func (w *Wallet) SignOut() error {
-	if err := w.touchDatastore(); err != nil {
-		return err
-	}
-	log.Debug("signing out...")
-
-	// remote is stateless, so we just ditch the local token
-	if err := w.datastore.Profile().SignOut(); err != nil {
-		log.Errorf("local signout error: %s", err)
-		return err
-	}
-	return nil
-}
-
-// IsSignedIn returns whether or not a user is signed in
-func (w *Wallet) IsSignedIn() (bool, error) {
-	if err := w.touchDatastore(); err != nil {
-		return false, err
-	}
-	_, err := w.datastore.Profile().GetUsername()
-	return err == nil, nil
-}
-
-// GetUsername returns the current user's username
-func (w *Wallet) GetUsername() (string, error) {
-	if err := w.touchDatastore(); err != nil {
-		return "", err
-	}
-	return w.datastore.Profile().GetUsername()
-}
-
 // GetId returns peer id
 func (w *Wallet) GetId() (string, error) {
 	if !w.started {
@@ -433,26 +339,6 @@ func (w *Wallet) GetPubKeyString() (string, error) {
 	return libp2pc.ConfigEncodeKey(pkb), nil
 }
 
-// GetAccessToken returns the current access_token (jwt) for central
-func (w *Wallet) GetAccessToken() (string, error) {
-	if err := w.touchDatastore(); err != nil {
-		return "", err
-	}
-	at, _, err := w.datastore.Profile().GetTokens()
-	if err != nil {
-		return "", err
-	}
-	return at, nil
-}
-
-func (w *Wallet) GetCentralAPI() string {
-	return w.centralAPI
-}
-
-func (w *Wallet) GetCentralUserAPI() string {
-	return fmt.Sprintf("%s/api/v1/users", w.centralAPI)
-}
-
 func (w *Wallet) Threads() []*thread.Thread {
 	return w.threads
 }
@@ -468,9 +354,6 @@ func (w *Wallet) GetThread(id string) (*int, *thread.Thread) {
 
 // AddThread adds a thread with a given name and secret key
 func (w *Wallet) AddThread(name string, secret libp2pc.PrivKey) (*thread.Thread, error) {
-	log.Debugf("adding a new thread: %s", name)
-
-	// index a new thread
 	skb, err := secret.Bytes()
 	if err != nil {
 		return nil, err
@@ -512,6 +395,8 @@ func (w *Wallet) AddThread(name string, secret libp2pc.PrivKey) (*thread.Thread,
 
 	// notify listeners
 	w.sendUpdate(Update{Id: thrd.Id, Name: thrd.Name, Type: ThreadAdded})
+
+	log.Debugf("added a new thread %s with name %s", thrd.Id, name)
 
 	return thrd, nil
 }
@@ -561,7 +446,7 @@ func (w *Wallet) RemoveThread(id string) (mh.Multihash, error) {
 	// notify listeners
 	w.sendUpdate(Update{Id: thrd.Id, Name: thrd.Name, Type: ThreadRemoved})
 
-	log.Infof("removed thread '%s'", id)
+	log.Infof("removed thread %s with name %s", id, thrd.Name)
 
 	return addr, nil
 }
